Project Updated
Xamarin Essential Like code access.
NotificationCenter.Current.NotificationTapped += OnLocalNotificationTapped;

to

NotificationCenter.NotificationTapped += OnLocalNotificationTapped;
