+++ Source/Plugin.LocalNotification/NotificationReceivedEventArgs.cs
