+++ Source/Plugin.LocalNotification/NotificationRequestBuilder.cs
@@ -166,7 +166,7 @@ namespace Plugin.LocalNotification
 		/// Creates the notification request
 		/// </summary>
 		/// <returns>The notification request</returns>
-		public NotificationRequest Create() => new NotificationRequest()
+		public NotificationRequest Create() => new()
 		{
 			Android = AndroidOptions,
 			iOS = iOSOptions,
