+++ Source/Plugin.LocalNotification/Platform/Droid/NotificationChannelGroupRequestBuilder.cs
@@ -21,7 +21,7 @@ namespace Plugin.LocalNotification.Platform.Droid
 		/// Builds the request to <see cref="NotificationChannelGroupRequest"/>
 		/// </summary>
 		/// <returns></returns>
-		public NotificationChannelGroupRequest Build() => new NotificationChannelGroupRequest(Group, Name);
+		public NotificationChannelGroupRequest Build() => new(Group, Name);
 
 		/// <summary>
 		/// Sets the Group
