+++ Source/Plugin.LocalNotification/Plugin.LocalNotification.csproj
@@ -1,76 +1,83 @@
 ﻿<Project Sdk="MSBuild.Sdk.Extras/3.0.22">
 
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;MonoAndroid10.0;Xamarin.iOS10</TargetFrameworks>
-    <Product>$(AssemblyName) ($(TargetFramework))</Product>
-    <DefaultLanguage>en-US</DefaultLanguage>
-    <AutoGenerateBindingRedirects>True</AutoGenerateBindingRedirects>
-    <NeutralLanguage>en</NeutralLanguage>
-    <LangVersion>latest</LangVersion>
-    <CopyLocalLockFileAssemblies>false</CopyLocalLockFileAssemblies>
+	<PropertyGroup>
+		<TargetFrameworks>netstandard2.0;netstandard2.1;MonoAndroid10.0;Xamarin.iOS10</TargetFrameworks>
+		<Product>$(AssemblyName) ($(TargetFramework))</Product>
+		<DefaultLanguage>en-US</DefaultLanguage>
+		<AutoGenerateBindingRedirects>True</AutoGenerateBindingRedirects>
+		<NeutralLanguage>en</NeutralLanguage>
+		<LangVersion>latest</LangVersion>
+		<CopyLocalLockFileAssemblies>false</CopyLocalLockFileAssemblies>
 
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <Authors>Elvin (Tharindu) Thudugala</Authors>
-    <PackageTags>android, ios, xamarin, xamarin.forms, local, notification, local.notification</PackageTags>
-    <PackageProjectUrl>https://github.com/thudugala/Plugin.LocalNotification</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/thudugala/Plugin.LocalNotification</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <Description>The local notification plugin provides a way to show local notifications from Xamarin apps.</Description>
-    <PackageIcon>icon.png</PackageIcon>
-    <Copyright>Copyright © Elvin (Tharindu) Thudugala</Copyright>
-    <Version>5.2.0</Version>
-    <PackageReleaseNotes>Check: https://github.com/thudugala/Plugin.LocalNotification/releases </PackageReleaseNotes>
+		<NoWarn>CS1591</NoWarn>
 
-    <PublishRepositoryUrl>true</PublishRepositoryUrl>
-    <EmbedUntrackedSources>true</EmbedUntrackedSources>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+		<PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+		<Authors>Elvin (Tharindu) Thudugala</Authors>
+		<PackageTags>android, ios, xamarin, xamarin.forms, local, notification, local.notification</PackageTags>
+		<PackageProjectUrl>https://github.com/thudugala/Plugin.LocalNotification</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/thudugala/Plugin.LocalNotification</RepositoryUrl>
+		<RepositoryType>git</RepositoryType>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<GenerateDocumentationFile>true</GenerateDocumentationFile>
+		<Description>The local notification plugin provides a way to show local notifications from Xamarin apps.</Description>
+		<PackageIcon>icon.png</PackageIcon>
+		<Copyright>Copyright © Elvin (Tharindu) Thudugala</Copyright>
+		<Version>5.2.0</Version>
+		<PackageReleaseNotes>Check: https://github.com/thudugala/Plugin.LocalNotification/releases </PackageReleaseNotes>
 
-    <CodeAnalysisRuleSet>Plugin.LocalNotification.ruleset</CodeAnalysisRuleSet>
-  </PropertyGroup>
+		<PublishRepositoryUrl>true</PublishRepositoryUrl>
+		<EmbedUntrackedSources>true</EmbedUntrackedSources>
+		<IncludeSymbols>true</IncludeSymbols>
+		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" Version="5.0.3">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-  </ItemGroup>
+		<CodeAnalysisRuleSet>Plugin.LocalNotification.ruleset</CodeAnalysisRuleSet>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <None Include="..\..\Screenshots\icon.png" Pack="true" PackagePath="" Link="Properties\icon.png" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" Version="5.0.3">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <Compile Remove="**/Platform/**/*.cs" />
-    <None Include="**/Platform/**/*.cs" />
-  </ItemGroup>
+	<ItemGroup>
+		<None Include="..\..\Screenshots\icon.png" Pack="true" PackagePath="" Link="Properties\icon.png" />
+	</ItemGroup>
 
-  <ItemGroup Condition="$(TargetFramework.StartsWith('MonoAndroid'))">
-    <PackageReference Include="Xamarin.AndroidX.Work.Runtime" Version="2.5.0.1" />
-    <PackageReference Include="Xamarin.AndroidX.Core" Version="1.3.2.3" />
-    <Compile Include="**/Platform/Droid/**/*.cs" />
-  </ItemGroup>
+	<ItemGroup>
+		<Compile Remove="**/Platform/**/*.cs" />
+		<None Include="**/Platform/**/*.cs" />
+	</ItemGroup>
 
-  <ItemGroup Condition="$(TargetFramework.StartsWith('Xamarin.iOS'))">
-    <Compile Include="**/Platform/iOS/**/*.cs" />
-  </ItemGroup>
+	<ItemGroup Condition="$(TargetFramework.StartsWith('MonoAndroid'))">
+		<PackageReference Include="Xamarin.AndroidX.Work.Runtime" Version="2.5.0.1" />
+		<PackageReference Include="Xamarin.AndroidX.Core" Version="1.3.2.3" />
+		<Compile Include="**/Platform/Droid/**/*.cs" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <Compile Update="Properties\Resources.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
-  </ItemGroup>
+	<ItemGroup Condition="$(TargetFramework.StartsWith('Xamarin.iOS'))">
+		<Compile Include="**/Platform/iOS/**/*.cs" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <EmbeddedResource Update="Properties\Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-  </ItemGroup>
+	<ItemGroup Condition="$(TargetFramework.StartsWith('netstandard'))">
+		<Compile Include="**/Platform/NetStandard/**/*.cs" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<Compile Update="Properties\Resources.Designer.cs">
+			<DesignTime>True</DesignTime>
+			<AutoGen>True</AutoGen>
+			<DependentUpon>Resources.resx</DependentUpon>
+		</Compile>
+	</ItemGroup>
+
+	<ItemGroup>
+		<EmbeddedResource Update="Properties\Resources.resx">
+			<Generator>ResXFileCodeGenerator</Generator>
+			<LastGenOutput>Resources.Designer.cs</LastGenOutput>
+		</EmbeddedResource>
+	</ItemGroup>
+
 </Project>
