+++ Source/Plugin.LocalNotification/iOSOptions.cs
